Remove simple_finalize, expose passes::run_basic
