+++ qirlib/src/lib.rs
@@ -17,7 +17,7 @@ pub mod builder;
 #[cfg(not(feature = "no-llvm-linking"))]
 pub mod evaluation;
 #[cfg(not(feature = "no-llvm-linking"))]
-pub mod module;
+pub mod passes;
 #[cfg(not(feature = "no-llvm-linking"))]
 pub mod qis;
 #[cfg(all(test, not(feature = "no-llvm-linking")))]
