+++ qirlib/src/tests.rs
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-use crate::{module, values::create_entry_point};
-use inkwell::{builder::Builder, context::Context};
+use crate::{passes, values::create_entry_point};
+use inkwell::{builder::Builder, context::Context, support::LLVMString};
 use normalize_line_endings::normalized;
 use std::{env, fs, path::PathBuf};
 
@@ -21,7 +21,8 @@ pub(crate) fn assert_reference_ir(
 ) -> Result<(), String> {
     const PYQIR_TEST_SAVE_REFERENCES: &str = "PYQIR_TEST_SAVE_REFERENCES";
     let (prefix, name) = split_id(id);
-    let actual_ir = build_ir(name, required_num_qubits, required_num_results, build)?;
+    let actual_ir = build_ir(name, required_num_qubits, required_num_results, build)
+        .map_err(|e| e.to_string())?;
 
     let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
     path.push("resources");
@@ -32,8 +33,7 @@ pub(crate) fn assert_reference_ir(
 
     if env::var(PYQIR_TEST_SAVE_REFERENCES).is_ok() {
         fs::create_dir_all(path.parent().unwrap()).map_err(|e| e.to_string())?;
-        fs::write(&path, actual_ir).map_err(|e| e.to_string())?;
-
+        fs::write(&path, actual_ir.to_bytes()).map_err(|e| e.to_string())?;
         Err(format!(
             "Saved reference IR. Run again without the {} environment variable.",
             PYQIR_TEST_SAVE_REFERENCES
@@ -41,7 +41,7 @@ pub(crate) fn assert_reference_ir(
     } else {
         let contents = fs::read_to_string(&path).map_err(|e| e.to_string())?;
         let expected_ir: String = normalized(contents.chars()).collect();
-        assert_eq!(expected_ir, actual_ir);
+        assert_eq!(expected_ir, actual_ir.to_str().map_err(|e| e.to_string())?);
         Ok(())
     }
 }
@@ -51,21 +51,24 @@ fn build_ir(
     required_num_qubits: u64,
     required_num_results: u64,
     build: impl for<'ctx> Fn(&Builder<'ctx>),
-) -> Result<String, String> {
+) -> Result<LLVMString, LLVMString> {
     let context = Context::create();
     let module = context.create_module(name);
     let entry_point =
         create_entry_point(&module, "main", required_num_qubits, required_num_results);
+
     let builder = context.create_builder();
     builder.position_at_end(context.append_basic_block(entry_point, ""));
     build(&builder);
     builder.build_return(None);
-    module::simple_finalize(&module)?;
-    Ok(module.print_to_string().to_string())
+
+    passes::run_basic(&module);
+    module.verify()?;
+    Ok(module.print_to_string())
 }
 
 fn split_id(id: &str) -> (Vec<&str>, &str) {
     let mut parts: Vec<_> = id.split('/').collect();
-    let name = parts.pop().unwrap();
+    let name = parts.pop().expect("Empty string.");
     (parts, name)
 }
